==> 2_cosine_similarity.ipynb <==
# 2. Cosine Similarity

# imports

import pandas as pd
import numpy as np

from sklearn.metrics.pairwise import cosine_similarity

# import the subset data

subset = pd.read_parquet('data/subset.parquet')

# view the subset data

subset

# calculate the user correlation using cosine similarity

user_sim = cosine_similarity(subset)

# create a dataframe with the user similarities

user_sim = pd.DataFrame(user_sim, columns=subset.index, index=subset.index)
user_sim

# save the user similarities to parquet

user_sim.to_parquet('data/subset_user_sim.parquet')

# create a dataframe with the item similarities

item_sim = cosine_similarity(subset.T)
item_sim = pd.DataFrame(item_sim, columns=subset.columns, index=subset.columns)
item_sim

# save the item similarities to parquet

item_sim.to_parquet('data/subset_item_sim.parquet')
